Move FMD code to new module
